Meta data visualization: Separate frontend (currently developed in angular). Add path to frontend in NomadMetaInfoServive and method to get complete data
A separate lightweight frontend code that uses the REST api of the backend to fetch and display data. The frontend is kept in the same as a part of the backend as backend is not configured to handle Cross-origin resource sharing (CORS), in the future when backend has been configured to handle CORS, frontend could be kept as a separate project.  Currently the frontend is in AngularJs but it will be changed in the future.

Add path (/api/index.html) to the frontend in the NomadMetaInfoService.scala. Add function to get complete data including ancestors and children.